+++ payment-service/src/main/java/com/devteria/payment/dto/request/BookingConfirmedEmailRequest.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -13,6 +15,6 @@ import lombok.experimental.FieldDefaults;
 
 public class BookingConfirmedEmailRequest {
     String bookingId;
-    Double finalPrice;
+    BigDecimal finalPrice;
     String voucherId;
 }
